+++ packages/maskbook/src/plugins/FileService/SNSAdaptor/components/Upload.tsx
@@ -7,7 +7,7 @@ import { Trans } from 'react-i18next'
 import { useHistory } from 'react-router'
 import { useAsync } from 'react-use'
 import { useI18N } from '../../../../utils'
-import { makeFileKey } from '../../makeFileKey'
+import { makeFileKey } from '../../file-key'
 import { FileRouter, MAX_FILE_SIZE } from '../../constants'
 import { PluginFileServiceRPC } from '../../Worker/rpc'
 import { RecentFiles } from './RecentFiles'
